just commit

+++ publisher.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import rospy
-from std_msgs.msg import Int32
+from std_msgs.msg import Int8
 from geometry_msgs.msg import Twist
 
 def in_range(x, min_, max_):
@@ -14,8 +14,8 @@ def move(twist: Twist):
     right_speed.publish(in_range(-int((twist.linear.x*2 + twist.angular.z)*speed), -100, 100))
 
 rospy.init_node('pc')
-left_speed = rospy.Publisher('speed_L', Int32, queue_size=10)
-right_speed = rospy.Publisher('speed_R', Int32, queue_size=10)
+left_speed = rospy.Publisher('speed_L', Int8, queue_size=10)
+right_speed = rospy.Publisher('speed_R', Int8, queue_size=10)
 keyboard_input = rospy.Subscriber('cmd_vel', Twist, move)
 
 speed = 20
